% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chi_sq_raw_material_by_phase}
\alias{chi_sq_raw_material_by_phase}
\title{Chi-sq for raw material by depth}
\usage{
chi_sq_raw_material_by_phase(plot_raw_materials_technology)
}
\arguments{
\item{plot_raw_materials_technology}{}
}
\description{
Chi-sq for raw material by depth
}
