% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_points_in_main_excavation_area}
\alias{clean_points_in_main_excavation_area}
\title{clean_points_in_main_excavation_area}
\usage{
clean_points_in_main_excavation_area(rotated_points_in_main_excavation_area)
}
\arguments{
\item{rotated_points_in_main_excavation_area}{}
}
\value{
a data frame
}
\description{
clean_points_in_main_excavation_area
}

