% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{combine_2012_and_2015_total_station_data}
\alias{combine_2012_and_2015_total_station_data}
\title{combine_2012_and_2015_total_station_data}
\usage{
combine_2012_and_2015_total_station_data(total_station_data_2012,
  total_station_data_2015)
}
\arguments{
\item{total_station_data_2015}{}
}
\value{
a data frame
}
\description{
combine_2012_and_2015_total_station_data
}

