% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_points_in_main_excavation_area}
\alias{extract_points_in_main_excavation_area}
\title{extract_points_in_main_excavation_area}
\usage{

  extract_points_in_main_excavation_area(combined_2012_and_2015_total_station_data)
}
\arguments{
\item{combined_2012_and_2015_total_station_data}{}
}
\value{
a data frame
}
\description{
extract_points_in_main_excavation_area
}

