% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{geoarchaeology}
\alias{geoarchaeology}
\title{geoarchaeology}
\usage{
geoarchaeology(cleaned_rotated_points_in_main_excavation_area)
}
\arguments{
\item{cleaned_rotated_points_in_main_excavation_area}{}
}
\value{
A data frame
}
\description{
geoarchaeology
}

