% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_c14_ages}
\alias{get_c14_ages}
\title{get_c14_ages}
\usage{
get_c14_ages(cleaned_rotated_points_in_main_excavation_area)
}
\arguments{
\item{cleaned_rotated_points_in_main_excavation_area}{}
}
\description{
get_c14_ages
}

