% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_osl_ages}
\alias{get_osl_ages}
\title{get_osl_ages}
\usage{
get_osl_ages(cleaned_rotated_points_in_main_excavation_area)
}
\arguments{
\item{cleaned_rotated_points_in_main_excavation_area}{}
}
\value{
a data frame
}
\description{
get_osl_ages
}

