% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{mjbnaturepaper}
\alias{mjbnaturepaper}
\alias{mjbnaturepaper-package}
\title{mjbnaturepaper: A package containing code and data for analysis of the 2012 and 2015 MJB excavations}
\description{
The foo package provides ...
}
\section{Foo functions}{

The foo functions ...
}

