% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ages_and_artefacts}
\alias{plot_ages_and_artefacts}
\title{plot_ages_and_artefacts}
\usage{
plot_ages_and_artefacts(osl_ages, c14_ages, stone_artefacts_only)
}
\arguments{
\item{stone_artefacts_only}{}
}
\description{
plot_ages_and_artefacts
}
