% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_geoarchaeology_data}
\alias{plot_geoarchaeology_data}
\title{plot_geoarchaeology_data}
\usage{
plot_geoarchaeology_data(prepared_geoarchaeology_data)
}
\arguments{
\item{prepared_geoarchaeology_data}{}
}
\value{
plots
}
\description{
plot_geoarchaeology_data
}

