% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_granulometry_data}
\alias{plot_granulometry_data}
\title{plot_granulometry_data}
\usage{
plot_granulometry_data(prepared_geoarchaeology_data)
}
\arguments{
\item{prepared_geoarchaeology_data}{}
}
\value{
plots
}
\description{
plot_granulometry_data
}
