% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_stone_artefacts}
\alias{plot_stone_artefacts}
\title{plot_stone_artefacts}
\usage{
plot_stone_artefacts(stone_artefacts_only)
}
\arguments{
\item{stone_artefacts_only}{}
}
\value{
A plot
}
\description{
plot_stone_artefacts
}

