% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prepare_geoarchaeology_data}
\alias{prepare_geoarchaeology_data}
\title{prepare_geoarchaeology_data}
\usage{
prepare_geoarchaeology_data(cleaned_rotated_points_in_main_excavation_area)
}
\arguments{
\item{cleaned_rotated_points_in_main_excavation_area}{}
}
\value{
A data frame
}
\description{
prepare_geoarchaeology_data
}

