% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{raw_materials_technology_plots}
\alias{raw_materials_technology_plots}
\title{raw material and technology plots}
\usage{
raw_materials_technology_plots(B6_raw_materials, spit_depths_B6_output)
}
\value{
plots
}
\description{
raw material and technology plots
}
