% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_2012_total_station_data}
\alias{read_2012_total_station_data}
\title{read_2012_total_station_data}
\usage{
read_2012_total_station_data(total_station_csv_directory_2012)
}
\arguments{
\item{total_station_csv_directory_2012}{A directory containing one or more CSV files}
}
\value{
a data frame
}
\description{
read_2012_total_station_data
}

