% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_non_artefacts}
\alias{remove_non_artefacts}
\title{remove non-artefacts}
\usage{
remove_non_artefacts(stone_artefacts_only)
}
\arguments{
\item{stone_artefacts_only}{}
}
\description{
remove non-artefacts
}
\examples{

}
