% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rotate_points_in_main_excavation_area}
\alias{rotate_points_in_main_excavation_area}
\title{rotate_points_in_main_excavation_area}
\usage{
rotate_points_in_main_excavation_area(extracted_points_in_main_excavation_area)
}
\arguments{
\item{extracted_points_in_main_excavation_area}{}
}
\value{
a data frame
}
\description{
rotate_points_in_main_excavation_area
}
