% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{size_sorting_stat_test}
\alias{size_sorting_stat_test}
\title{size sorting stat test}
\usage{
size_sorting_stat_test(spit_depths_B6_output, the_n = 10,
  p_value = 0.05)
}
\value{
a data frame
}
\description{
size sorting stat test
}
