% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stone_artefacts}
\alias{stone_artefacts}
\title{stone_artefacts}
\usage{
stone_artefacts(cleaned_rotated_points_in_main_excavation_area)
}
\arguments{
\item{cleaned_rotated_points_in_main_excavation_area}{}
}
\value{
A data frame
}
\description{
stone_artefacts
}

